Better detection of failed network load.
When page load fails per QWebView.onLoadFinished, check to see if the
requested URL was one of the URLs that failed or if it was some
nonessential url.

This fixes a bug where pages don't load when some nonessential
resource falsely triggers a fail condition, even though the page
itself would otherwise load.